iikoBiz resp customer

+++ OzonCard.BizClient/Models/Walletbalance.cs
@@ -3,7 +3,7 @@ namespace OzonCard.BizClient.Models
 {
     public class Walletbalance
     {
-        public decimal balance { get; set; }
+        public double balance { get; set; }
         public Wallet wallet { get; set; }
     }
